Implement serde support for `aws_smithy_types::Number` (#2185)
